First major refactoring of moving the meat of the simulation code into a function.
This version of the function takes two different sets of default
parameters depending on whether or not the chaos=True flag is used.

This still otherwise maintains the existing structure of the file, in
that it is a script which can be run directly and that returns two
plots.  To this end we call the function in the body of the module, and
use the return values to generate the two plots.